created parent class for LiveChatTest
