+++ src/test/java/LiveChatAPI/LiveChatTest/GetIntitalDataTest.java
@@ -1,5 +1,6 @@
 package LiveChatAPI.LiveChatTest;
 
+import LiveChatAPI.LiveChatTest.LiveChatParent.ChatParent;
 import io.rocketchat.common.data.model.ErrorObject;
 import io.rocketchat.livechat.LiveChatAPI;
 import io.rocketchat.livechat.callback.ConnectListener;
@@ -15,11 +16,7 @@ import static org.mockito.Mockito.timeout;
 /**
  * Created by sachin on 17/7/17.
  */
-public class GetIntitalDataTest implements ConnectListener {
-
-    private static String serverurl="wss://livechattest.rocket.chat/websocket";
-
-    LiveChatAPI api;
+public class GetIntitalDataTest extends ChatParent implements ConnectListener {
 
     @Mock
     InitialDataListener dataListener;
@@ -31,11 +28,13 @@ public class GetIntitalDataTest implements ConnectListener {
     ArgumentCaptor <ErrorObject> error;
 
     @Before
-    public void setUpBefore(){
-        MockitoAnnotations.initMocks( this );
-        System.out.println("before got called");
-        api= new LiveChatAPI(serverurl);
-        api.setReconnectionStrategy(null);
+    public void setup(){
+        setUpBefore();
+    }
+
+    @Override
+    public void setUpBefore() {
+        super.setUpBefore();
         api.connect(this);
     }
